Rename controls with a name instead of 1 2 3
